acomodo archivos del back ordenado ruta para agregar avatar, ahora solo ingresa el avatar a la tabla user sin tocar infouser
