+++ api/src/controlers/users.js
@@ -12,6 +12,9 @@ const Address = require('../models/Address');
 async function findUser(userName) {
     try {
         const response = await User.findOne(userName)
+        .populate('avatar')
+        .populate('info')
+        .populate('partner')
         return response
     } catch (error) {
         console.log(error.message)
@@ -32,48 +35,84 @@ async function findAllUsers() {
     }
 }
 
-async function createUser(newUser) {
+const getUser = async (req, res) => {
+    const { id } = req.params;
+    console.log(id)
     try {
-        const response = await User.create({
-            userName: newUser.username,
-            password: newUser.password,
-            latitude: newUser.latitude,
-            longitude: newUser.longitude,
-            type: newUser.type,
+        const user = await User.findById(id)
+            .populate('avatar')
+            .populate('info')
+            .populate('partner')
+        console.log(user)
+        res.json({
+            ok: true,
+            user
         })
-        return response
     } catch (error) {
-        console.log(error.message)
-        return error.message
+        console.log(error)
+        res.status(500).json({
+            ok: false,
+            msg: "Unexpected error"
+        })
     }
 }
 
-const updateAvatarForUser = async (req, res) => {
-    const { id } = req.params;
+const updateUser = async (req, res) => {
+    const { id } = req.params
     try {
-        const UserUpdateAvatar = await User.findByIdAndUpdate(
-            id,
-            req.body,
-            { new: true }
-        )
-        const idForInfo = UserUpdateAvatar.info
-        const UserInfoUpdateAvatar = await InfoUser.findByIdAndUpdate(
-            idForInfo,
-            req.body,
-            { new: true }
-        )
+        const body = req.body
+        const newAddressUser = {
+            street: body.street,
+            floor: body.floor,
+            address: body.address,
+            apartament: body.apartament,
+            neighborhood: body.neighborhood,
+            city: body.city,
+            country: body.country,
+            zipCode: body.zipCode
+        }
+        const user = await User.findById(id)
+        const idAddress = user.address ? user.address : null;
+        if (idAddress === null) {
+            const addressUser = new Address(newAddressUser)
+            await addressUser.save()
+            idAddress = addressUser._id
+        } else {
+            let updatedAddress = await Address.findByIdAndUpdate(idAddress, newAddressUser, { new: true })
+        }        
+        const idInfo = user.info
+        const idAvatar = user.avatar
+        const newInfoUser = {
+            username: body.username,
+            lastName: body.lastname,
+            phone: body.phone,
+            birthday: body.birthday,
+            avatar: idAvatar,
+            address: idAddress,
+            gender: body.gender,
+            photo: body.photo,
+        }
+        const updUser = await InfoUser.findByIdAndUpdate(idInfo, newInfoUser, { new: true })
         res.status(200).json({
             ok: true,
-            msg: "Usuario modificado correctamente",
-            UserUpdateAvatar,
-            UserInfoUpdateAvatar
+            updUser
         })
     } catch (error) {
+        console.log(error)
         res.status(500).json({
             ok: false,
-            msg: "No se pudo modificar el usuario"
+            msg: "no se pudo actualizar el usuario"
         })
-        console.log("error: ", error)
+    }
+}
+
+async function deleteUser(id) {
+    try {
+        const userDeleted = await User.findByIdAndDelete(id)
+        console.log(userDeleted)
+    } catch (error) {
+        console.log(error.message)
+        return error.message
     }
 }
 
@@ -96,16 +135,6 @@ const getUserGoogleAccount = async (req, res) => {
             ok: false,
             msg: "Unexpected errorf"
         })
-    }
-}
-
-async function deleteUser(id) {
-    try {
-        const userDeleted = await User.findByIdAndDelete(id)
-        console.log(userDeleted)
-    } catch (error) {
-        console.log(error.message)
-        return error.message
     }
 }
 
@@ -150,28 +179,6 @@ const googleSignIn = async (req, res) => {
         res.status(500).json({
             ok: false,
             msg: "No se pudo crear el usuario"
-        })
-    }
-}
-
-const getUser = async (req, res) => {
-    const { id } = req.params;
-    console.log(id)
-    try {
-        const user = await User.findById(id)
-            .populate('avatar')
-            .populate('info')
-            .populate('partner')
-        console.log(user)
-        res.json({
-            ok: true,
-            user
-        })
-    } catch (error) {
-        console.log(error)
-        res.status(500).json({
-            ok: false,
-            msg: "Unexpected error"
         })
     }
 }
@@ -265,50 +272,15 @@ async function updatePassword(userId, newPassword, password, secretToken) {
     }
 }
 
-const updateUser = async (req, res) => {
-    const { id } = req.params
-    try {
-        const body = req.body
-        const user = await User.findById(id)
-        const addressUser = new Address({
-            street: body.street,
-            floor: body.floor,
-            address: body.address,
-            apartament: body.apartament,
-            neighborhood: body.neighborhood,
-            city: body.city,
-            country: body.country,
-            zipCode: body.zipCode
-        })
-        await addressUser.save()
-        const idAddress = addressUser._id
-        const idInfo = user.info
-        const idAvatar = user.avatar
-        const newInfoUser = {
-            username: body.username,
-            lastName: body.lastname,
-            phone: body.phone,
-            birthday: body.birthday,
-            avatar: idAvatar,
-            address: idAddress,
-            gender: body.gender,
-            photo: body.photo,
-        }
-        const updUser = await InfoUser.findByIdAndUpdate(idInfo, newInfoUser, { new: true })
-        res.status(200).json({
-            ok: true,
-            updUser
-        })
-    } catch (error) {
-        console.log(error)
-        res.status(500).json({
-            ok: false,
-            msg: "no se pudo actualizar el usuario"
-        })
-    }
-}
 
-
-
-
-module.exports = { findUser, findAllUsers, createUser, deleteUser, updateAvatarForUser, googleSignIn, getUser, isValidObjectId, updatePassword, getUserGoogleAccount, updateUser }
+module.exports = {
+  findUser,
+  findAllUsers,
+  getUser,
+  deleteUser,
+  updateUser,
+  updatePassword,
+  googleSignIn,
+  isValidObjectId,
+  getUserGoogleAccount,
+};
